update -- authentication wallet & account & jwt token

+++ plutocart_backend/src/main/java/com/example/plutocart/services/AccountService.java
@@ -1,5 +1,6 @@
 package com.example.plutocart.services;
 
+import com.example.plutocart.auth.JwtUtil;
 import com.example.plutocart.dtos.account.AccountDTO;
 import com.example.plutocart.entities.Account;
 import com.example.plutocart.repositories.AccountRepository;
@@ -22,21 +23,24 @@ public class AccountService {
     public GenericResponse CreateAccountByImei(AccountDTO accountDTO){
         IMEIEncryption imeiEncryption = new IMEIEncryption();
         GenericResponse response = new GenericResponse();
-       try {
-           String imE = imeiEncryption.encryptIMEI(accountDTO.getImei());
-           accountRepository.CreateAccountByImei(imE);
-           Account account = accountRepository.getAccountByImeiAndRole(imE , 1);
-           accountDTO.setImei(imE);
-           accountDTO.setAccountRole(account.getAccountRole());
-           accountDTO.setAccountId(account.getAccountId());
-           response.setStatus(ResultCode.SUCCESS);
-           response.setData(accountDTO);
-           return response;
-       }
-       catch (Exception ex){
-           response.setStatus(ResultCode.BAD_REQUEST);
-       }
-       return response;
+        try {
+            String imE = imeiEncryption.encryptIMEI(accountDTO.getImei());
+            accountRepository.CreateAccountByImei(imE);
+            Account account = accountRepository.getAccountByImeiAndRole(imE , 1);
+            accountDTO.setImei(imE);
+            accountDTO.setAccountRole(account.getAccountRole());
+            accountDTO.setAccountId(account.getAccountId());
+            response.setStatus(ResultCode.SUCCESS);
+            response.setData(accountDTO);
+            String acId = String.valueOf(accountDTO.getAccountId());
+            String token = JwtUtil.generateToken(acId);
+            response.setAuthentication(token);
+            return response;
+        }
+        catch (Exception ex){
+            response.setStatus(ResultCode.BAD_REQUEST);
+        }
+        return response;
     }
 
     public GenericResponse CreateAccountByGoogle(AccountDTO accountDTO){
@@ -53,6 +57,9 @@ public class AccountService {
                 accountDTO.setAccountId(account.getAccountId());
                 response.setStatus(ResultCode.SUCCESS);
                 response.setData(accountDTO);
+                String acId = String.valueOf(accountDTO.getAccountId());
+                String token = JwtUtil.generateToken(acId);
+                response.setAuthentication(token);
                 return response;
             }
 
@@ -63,39 +70,54 @@ public class AccountService {
         return response;
     }
 
-    public GenericResponse UpdateAccountToMember(String email , Integer accountId){
+    public GenericResponse UpdateAccountToMember(String email , Integer accountId , String token){
         GenericResponse response = new GenericResponse();
-        try {
-            accountRepository.updateAccountToMember(email , accountId);
-            response.setStatus(ResultCode.SUCCESS);
-            Account account = accountRepository.findById(accountId).get();
-            account.setAccountRole("Member");
-            account.setEmail(email);
-            response.setData(account);
-            return response;
+        String userId = JwtUtil.extractUsername(token);
+        if(Integer.parseInt(userId) == accountId){
+            try {
+                accountRepository.updateAccountToMember(email , accountId);
+                response.setStatus(ResultCode.SUCCESS);
+                Account account = accountRepository.findById(accountId).get();
+                account.setAccountRole("Member");
+                account.setEmail(email);
+                response.setData(account);
+                return response;
+            }
+            catch (Exception ex){
+                response.setStatus(ResultCode.BAD_REQUEST);
+                response.setData(null);
+                return response;
+            }
         }
-        catch (Exception ex){
-            response.setStatus(ResultCode.BAD_REQUEST);
+        else {
+            response.setStatus(ResultCode.FORBIDDEN);
             response.setData(null);
             return response;
         }
+
     }
 
-    public GenericResponse DeleteAccount(Integer accountId){
+    public GenericResponse DeleteAccount(Integer accountId , String token){
         GenericResponse response = new GenericResponse();
-        try {
-            accountRepository.deleteAccount(accountId);
-            response.setStatus(ResultCode.SUCCESS);
-            return response;
+        String userId = JwtUtil.extractUsername(token);
+        System.out.println("using DeleteAccount Service");
+        if(Integer.parseInt(userId) == accountId){
+            try {
+                accountRepository.deleteAccount(accountId);
+                response.setStatus(ResultCode.SUCCESS);
+                return response;
+            }
+            catch (Exception ex){
+                response.setStatus(ResultCode.BAD_REQUEST);
+                response.setData(null);
+                return response;
+            }
         }
-        catch (Exception ex){
-            response.setStatus(ResultCode.BAD_REQUEST);
+        else {
+            response.setStatus(ResultCode.FORBIDDEN);
             response.setData(null);
-            return response;
+            return  response;
         }
     }
-
-
-
 
 }
